skip yarn cleanup by default

+++ build.rs
@@ -4,6 +4,7 @@ use std::process::Command;
 
 fn main() {
     let skip_yarn = env::var("SKIP_YARN").unwrap_or_default() == "true";
+    let skip_yarn_cleanup = env::var("SKIP_YARN_CLEANUP").unwrap_or(String::from("true")) == "true";
     if !skip_yarn {
         Command::new("yarn")
             .arg("install")
@@ -14,7 +15,9 @@ fn main() {
             .arg("build")
             .output()
             .expect("failed to run yarn build");
-        fs::remove_dir_all(".svelte-kit").expect("failed to delete .svelte-kit");
-        fs::remove_dir_all("node_modules").expect("failed to delete node_modules");
+        if !skip_yarn_cleanup {
+            fs::remove_dir_all(".svelte-kit").expect("failed to delete .svelte-kit");
+            fs::remove_dir_all("node_modules").expect("failed to delete node_modules");
+        }
     }
 }
